requirements.txt
openai
pydantic
python-dotenv
pytest
httpx
pytest-asyncio
pytest-json-report
pytest-html
ragas
playwright
